+++ metrics/eye_storage.cpp
@@ -37,7 +37,7 @@ bool Load(std::string const & filename, std::vector<int8_t> & dst)
     FileReader reader(filename);
 
     dst.clear();
-    dst.resize(reader.Size());
+    dst.resize(static_cast<size_t>(reader.Size()));
 
     reader.Read(0, dst.data(), dst.size());
   }
